Updates from comments
- change from Long to bytes[] in proto
- JVM methods return BigInteger from now
- in Python conversion from BigInteger to int is via bytes[]
- drop .dir-locals.el from .gitignore
- rename _sizeInBytes -> _size_in_bytes on Python side

 On branch size_in_bytes_api
 Your branch is up to date with 'origin/size_in_bytes_api'.

 Changes to be committed:
	modified:   .gitignore
	modified:   connector/connect/client/jvm/src/main/scala/org/apache/spark/sql/Dataset.scala
	modified:   connector/connect/client/jvm/src/main/scala/org/apache/spark/sql/SparkSession.scala
	modified:   connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/ClientE2ETestSuite.scala
	modified:   connector/connect/common/src/main/protobuf/spark/connect/base.proto
	modified:   connector/connect/server/src/main/scala/org/apache/spark/sql/connect/service/SparkConnectAnalyzeHandler.scala
	modified:   python/pyspark/sql/classic/dataframe.py
	modified:   python/pyspark/sql/connect/client/core.py
	modified:   python/pyspark/sql/connect/dataframe.py
	modified:   python/pyspark/sql/connect/proto/base_pb2.py
	modified:   python/pyspark/sql/connect/proto/base_pb2.pyi
	modified:   python/pyspark/sql/dataframe.py
	modified:   sql/core/src/main/scala/org/apache/spark/sql/Dataset.scala

+++ connector/connect/client/jvm/src/main/scala/org/apache/spark/sql/Dataset.scala
@@ -17,13 +17,11 @@
 package org.apache.spark.sql
 
 import java.util.{Collections, Locale}
-
 import scala.collection.mutable
 import scala.jdk.CollectionConverters._
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.TypeTag
 import scala.util.control.NonFatal
-
 import org.apache.spark.SparkException
 import org.apache.spark.annotation.DeveloperApi
 import org.apache.spark.api.java.function._
@@ -41,6 +39,8 @@ import org.apache.spark.sql.types.{Metadata, StructType}
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.util.ArrayImplicits._
 import org.apache.spark.util.SparkClassUtils
+
+import java.math.BigInteger
 
 /**
  * A Dataset is a strongly typed collection of domain-specific objects that can be transformed in
@@ -289,7 +289,7 @@ class Dataset[T] private[sql] (
    * @group basic
    * @since 4.0.0
    */
-  def sizeInBytesApproximation(): Long = {
+  def sizeInBytesApproximation(): BigInteger = {
     sparkSession.sizeInBytes(plan.getRoot)
   }
